refactor: split chat into codeeditor and chatmessages
