+++ src/components/chat/chat.js
@@ -1,23 +1,19 @@
-import React from 'react';
+import React, { Component } from 'react';
 import './chat.scss';
 
-import CodeMirror from 'codemirror';
-import 'codemirror/addon/edit/matchbrackets';
-import 'codemirror/mode/javascript/javascript';
-import 'codemirror/lib/codemirror.css';
+import CodeEditor from './codeeditor';
+import ChatMessages from './chat-messages';
+
 import Overlay from './overlay';
 import ModalEndChat from './../modal/modal-end-chat';
 
-class Chat extends React.Component {
-
-  textArea = React.createRef();
+class Chat extends Component {
 
   state = {
-    keepChangeEditor: '',
     roomId: this.props.location.pathname.split('/')[2],
     questionId: this.props.location.pathname.split('/')[3],
     tutorOrLearner: this.props.location.pathname.split('/')[4],
-    showModal: false,
+    showFeedbackModal: false,
     tutorJoined: false,
     minutes: 0,
     seconds: 0,
@@ -26,9 +22,6 @@ class Chat extends React.Component {
   }
 
   componentDidMount() {
-
-    // this.props.socket.on('join room', () => this.setState({tutorJoined: true}, () => this.startTimer()));
-
     //const room = this.props.room; //Amber removed this ... TTD to refractor
     this.props.socket.emit('join room', this.state.roomId)
 
@@ -37,126 +30,8 @@ class Chat extends React.Component {
       else this.setState({tutorJoined: false});
     });
 
-    // CHAT
-    this.props.socket.on('chat message', (msg) => {
-      let currentId = this.props.socket.id;
-      let oneMessage = document.createElement('div');
-      let messages = document.getElementById('messages').appendChild(oneMessage);
-      oneMessage.className += ' oneMessage';
-      let messageText = document.createElement('div');
-      let messageDate = document.createElement('div');
-      oneMessage.appendChild(messageText);
-      messageText.innerHTML = msg.value;
-      oneMessage.appendChild(messageDate);
-      messageDate.innerHTML = msg.date;
-      messageDate.className += ' messageDate';
-      if (currentId === msg.id) messages.className += ' myMessage';
-      else messages.className += ' elseMessage';
-    });
-
-    // CODE EDITOR
-    this.codemirror =  CodeMirror.fromTextArea(this.textArea.current, {
-      mode: "javascript",
-      theme: "default",
-      lineNumbers: true,
-      content: this.textArea.current,
-    })
-    this.codemirror.setSize(null, '80vh');
-    this.codemirror.on('blur', this.codeChanged);
-    this.props.socket.on('editor', (data) => this.codemirror.getDoc().setValue(data.code)); // handles received text
-    this.props.socket.on('newUser', this.updateCode); // this code is not working - what was its purpose?
-
     //HANG-UP
-    this.props.socket.on('hang up', () => {
-      this.openChatModal()
-    })
-  }
-
-  //WHY WAS THIS HERE? -- This was added by Arol to prevent the editor from continuously re-rendering
-  // shouldComponentUpdate() {
-  //   return false;
-  // }
-
-  updateCode = (data) => {
-    this.codemirror.getDoc().setValue(data.code);
-    this.setState({keepChangeEditor: this.codemirror.getDoc().getValue()})
-  }
-
-  clickButton = (e) => {
-    e.preventDefault();
-    this.sendMessage();
-  }
-
-  detectEnter = (e) => {
-    var code = e.keyCode || e.which;
-    if (code === 13) {
-      e.preventDefault();
-      this.sendMessage();
-    }
-  }
-
-  sendMessage = () => {
-    const dateNow = new Date();
-    const dateNowFormatted = dateNow.toLocaleString();
-    const msgToSend = {
-      date: dateNowFormatted,
-      id: this.props.socket.id,
-      value: this.message.value,
-      room: this.state.roomId
-    };
-    this.props.socket.emit('chat message', msgToSend);
-    this.message.value = '';
-  }
-
-  codeChanged = () => {
-    const editorContent = this.codemirror.getDoc().getValue();
-    const data = { code: editorContent, room: this.state.roomId } // changed property 'text' to 'code' to be more explicit
-    if (editorContent !== this.state.keepChangeEditor) {
-      this.props.socket.emit('editor', data);
-      this.setState({keepChangeEditor: this.codemirror.getDoc().getValue()});
-    }
-    // todo: update the chatmirror so it updates during typing, not just on clicking outside of the box
-    // var myElement = document.getElementById('txtArea');
-    // myElement.focus();
-    // var startPosition = myElement.selectionStart;
-    // var endPosition = myElement.selectionEnd;
-    // console.log('startPosition ', startPosition)
-    // console.log('endPosition ', endPosition)
-  }
-    
-  renderOverlay = () => {
-    if (this.state.tutorOrLearner === 'learner' && !this.state.tutorJoined) {
-      return <Overlay closeOverlay={(counter) => {
-        clearInterval(counter);
-        this.props.history.goBack()
-      }
-      }/>
-    }
-  }
-
-  hangUp = () => {
-    this.props.socket.emit('hang up', {roomId: this.state.roomId});
-    this.setState({
-      showModal: true
-    })
-  }
-
-  showEndChatModal = () => {
-    if (this.state.showModal) {
-      return <ModalEndChat closeChatModal={this.closeChatModal} history={this.props.history} questionId={this.state.questionId} tutorOrLearner={this.state.tutorOrLearner}/>
-    }
-  }
-
-  openChatModal = () => {
-    this.setState({
-      showModal: true
-    })
-  }
-
-  closeChatModal = () => {
-    this.setState({
-      showModal: false
-    })
+    this.props.socket.on('hang up', () => {this.setState({showFeedbackModal: true})})
   }
 
   startTimer = () => {
@@ -175,7 +50,26 @@ class Chat extends React.Component {
   
     }, 1000)
   }
+    
+  renderOverlay = () => {
+    if (this.state.tutorOrLearner === 'learner' && !this.state.tutorJoined) {
+      return <Overlay closeOverlay={(counter) => {
+        clearInterval(counter);
+        this.props.history.goBack()
+      }
+      }/>
+    }
+  }
 
+  hangUp = () => {
+    this.props.socket.emit('hang up', {roomId: this.state.roomId});
+  }
+
+  showEndChatModal = () => {
+    if (this.state.showFeedbackModal) {
+      return <ModalEndChat closeChatModal={() => this.setState({showFeedbackModal: false})} history={this.props.history} questionId={this.state.questionId} tutorOrLearner={this.state.tutorOrLearner}/>
+    }
+  }
 
   render() {
     // Notify that user is online (since navbar is not render)
@@ -188,33 +82,22 @@ class Chat extends React.Component {
 
         <div className="chat-header">
           <h1>Question Title</h1>
+
           <h3 id="timer" style={{color: this.state.overTime}}>{this.state.minutes}:{this.state.secondsString}</h3>
+
           <button onClick={this.hangUp}>End Call</button>
         </div>
+
         <div className="chat-body">
-          <div className="editor">
-            <textarea id="txtArea" name="txtArea" ref={this.textArea}/>
-          </div>
-          <div className="chat-box">
-            <div id="messages"></div>
-            <form action="">
-              <input
-                autoComplete="off"
-                ref={input => this.message = input}
-                onKeyPress={this.detectEnter}
-              />
-              <button onClick={this.clickButton} type="button">
-                Send
-              </button>
-            </form>
-          </div>
+          <CodeEditor socket={this.props.socket} room={this.state.roomId}/>
+          <ChatMessages socket={this.props.socket} room={this.state.roomId} />
         </div>
+
         {this.showEndChatModal()}
+
       </div>
     )
   }
 }
-
-
 
 export default Chat;
