+++ src/components/modal/modal-tutor-notification.js
@@ -3,7 +3,7 @@ import '../log-in/log-in.scss';
 import { Link } from "react-router-dom";
 import '../../phone.scss';
 
-class Modal extends Component {
+class TutorInvitation extends Component {
 
   render() {
     console.log()
@@ -12,7 +12,7 @@ class Modal extends Component {
         <div className="modal">
           <h1>HEYYY</h1>
           <p>Modal Body</p>
-          <Link to={`/chat/${this.props.question.room_id}/${this.props.question.question_id}/tutor`}>Modal Button</Link>
+          <Link onClick={this.props.enterChatroom} to={`/chat/${this.props.question.room_id}/${this.props.question.question_id}/tutor`}>Modal Button</Link>
           <i className="Phone is-animating"></i>
         </div>
       </div>
@@ -20,4 +20,4 @@ class Modal extends Component {
   }
 }
 
-export default Modal;
+export default TutorInvitation;
